code refactoring
refactored code, so that the VersionSaveComponent.tsx now
 contains the whole functionality and is no longer split
 over multiple components. Also adjusted some comments
